(PR changes): models' fields improvements
Use a more consice way to expand CORPORATIONS and MAPS for __in Query.

Increase the max_length of game_map and corporation so that they are
base 2 and do not waste byte space.

Prefech game object for __str__ method so that db is queried once when
printing multiple properties.

create migrations for the changes.