optimize and clean up data_importer.py
replace the `remap_values` function with dedicated mini functions
(`game_map_map` and `corporation_map`) to make the code cleaner and
increase performance.

Reposition functions and global variables in a python style.

Change the way IntegrityError is handled: add an error message telling
the user that the import has failed instead of pretending that it
succeeded.